remove unnecessary casts

+++ bee-message/src/unlock/unlock_blocks.rs
@@ -21,8 +21,8 @@ use core::{
     ops::Deref,
 };
 
-const PREFIXED_UNLOCK_BLOCKS_LENGTH_MIN: u16 = *UNLOCK_BLOCK_COUNT_RANGE.start() as u16;
-const PREFIXED_UNLOCK_BLOCKS_LENGTH_MAX: u16 = *UNLOCK_BLOCK_COUNT_RANGE.end() as u16;
+const PREFIXED_UNLOCK_BLOCKS_LENGTH_MIN: u16 = *UNLOCK_BLOCK_COUNT_RANGE.start();
+const PREFIXED_UNLOCK_BLOCKS_LENGTH_MAX: u16 = *UNLOCK_BLOCK_COUNT_RANGE.end();
 
 /// Error encountered while packing [`UnlockBlocks`].
 #[derive(Debug)]
